+++ src/web/app/static/assets/cricket_icon.ico
