+++ src/web/app/static/assets/flags/Afghanistan.jpg
