+++ src/web/app/static/assets/flags/Australia.jpg
