+++ src/web/app/static/assets/flags/Bermuda.jpg
