+++ src/web/app/static/assets/flags/East Africa.jpg
