+++ src/web/app/static/assets/flags/India.jpg
