+++ src/web/app/static/assets/flags/Ireland.jpg
