+++ src/web/app/static/assets/flags/Kenya.jpg
