+++ src/web/app/static/assets/flags/Namibia.jpg
