+++ src/web/app/static/assets/flags/New Zealand.jpg
