+++ src/web/app/static/assets/flags/Scotland.jpg
