+++ src/web/app/static/assets/flags/Sri Lanka.jpg
