+++ src/web/app/static/assets/flags/West Indies.jpg
