+++ src/web/app/static/js/bootstrap.js
